Improve mocking online service
